refactor: lottery event 고정 데이터에 대한 상수 파일 분리
