+++ client/src/constants/Main/lotteryEventData.ts
@@ -1,0 +1,34 @@
+import { LotteryEventProps } from "@/components/LotteryEvent";
+
+export const lotteryEventData: LotteryEventProps[] = [
+    {
+        rank: 1,
+        image: "/public/assets/main/lottery/prize-1.png",
+        prizeName: "CASPER Electric",
+        winnerCount: 1,
+    },
+    {
+        rank: 2,
+        image: "/public/assets/main/lottery/prize-2.png",
+        prizeName: "프라이탁 메신저백",
+        winnerCount: 5,
+    },
+    {
+        rank: 3,
+        image: "/public/assets/main/lottery/prize-3.png",
+        prizeName: "나이키 에어포스",
+        winnerCount: 5,
+    },
+    {
+        rank: 4,
+        image: "/public/assets/main/lottery/prize-4.png",
+        prizeName: "스탠리 텀블러",
+        winnerCount: 10,
+    },
+    {
+        rank: 5,
+        image: "/public/assets/main/lottery/prize-5.png",
+        prizeName: "스타벅스 커피 기프티콘",
+        winnerCount: 315,
+    },
+];
